Added the Luhn_algo code
This algorithm can verify if a card number is valid or not. Learned here the use of str.maketrans() function that helps to modify all occurences of a char in a string with another .
Also learned the use of translate to apply the maketrans modification to a string.
Finally learned the str[start:stop:step] that helps extract a portion from a string while precising the start , stop and step values.